# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SignupRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  deliveryChannel?: string | null;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type MessageFrequency =
  (typeof MessageFrequency)[keyof typeof MessageFrequency];

export const MessageFrequency = {
  yearly: "yearly",
  biannual: "biannual",
} as const;

export interface Message {
  id: number;
  userId: string;
  content: string;
  frequency: MessageFrequency;
  createdAt: string;
  deliverAt: string;
  isDelivered: boolean;
  isRead: boolean;
}

export type CreateMessageRequestFrequency =
  (typeof CreateMessageRequestFrequency)[keyof typeof CreateMessageRequestFrequency];

export const CreateMessageRequestFrequency = {
  yearly: "yearly",
  biannual: "biannual",
} as const;

export interface CreateMessageRequest {
  /**
   * @minLength 1
   * @maxLength 5000
   */
  content: string;
  frequency: CreateMessageRequestFrequency;
}

export interface MessagesResponse {
  messages: Message[];
}

export interface MessageResponse {
  message: Message;
}

export const DeleteMessageResponseValue = {
  success: true,
} as const;
export type DeleteMessageResponse = typeof DeleteMessageResponseValue;

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;