# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new account
 */
export const signupBodyPasswordMin = 6;

export const SignupBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(signupBodyPasswordMin),
  firstName: zod.string().optional(),
});

/**
 * @summary Log in with email and password
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.string(),
    email: zod.string().email().nullable(),
    firstName: zod.string().nullable(),
    lastName: zod.string().nullable(),
    profileImageUrl: zod.string().nullable(),
    phoneNumber: zod.string().nullish(),
    deliveryChannel: zod.string().nullish(),
  }),
});

/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().email().nullable(),
      firstName: zod.string().nullable(),
      lastName: zod.string().nullable(),
      profileImageUrl: zod.string().nullable(),
      phoneNumber: zod.string().nullish(),
      deliveryChannel: zod.string().nullish(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary End the current session
 */
export const LogoutHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get all messages for the current user
 */
export const GetMessagesHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const GetMessagesResponse = zod.object({
  messages: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      content: zod.string(),
      frequency: zod.enum(["yearly", "biannual"]),
      createdAt: zod.date(),
      deliverAt: zod.date(),
      isDelivered: zod.boolean(),
      isRead: zod.boolean(),
    }),
  ),
});

/**
 * @summary Create a new future message
 */
export const CreateMessageHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const createMessageBodyContentMax = 5000;

export const CreateMessageBody = zod.object({
  content: zod.string().min(1).max(createMessageBodyContentMax),
  frequency: zod.enum(["yearly", "biannual"]),
});

/**
 * @summary Delete a message
 */
export const DeleteMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteMessageHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const DeleteMessageResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Mark a delivered message as read
 */
export const MarkMessageReadParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkMessageReadHeader = zod.object({
  Authorization: zod
    .string()
    .optional()
    .describe("Opaque session token — `Bearer <sid>`."),
});

export const MarkMessageReadResponse = zod.object({
  message: zod.object({
    id: zod.number(),
    userId: zod.string(),
    content: zod.string(),
    frequency: zod.enum(["yearly", "biannual"]),
    createdAt: zod.date(),
    deliverAt: zod.date(),
    isDelivered: zod.boolean(),
    isRead: zod.boolean(),
  }),
});