# lib/api-zod/src/generated/types/signupRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignupRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  firstName?: string;
}